python: Fix return temporary variable in Number_AsLong()
This is cosmetic, since the value is cast to a long by the return
statement, but one may wonder what is the trick with using an
unsigned long long as a temporary... Well, there is none, it was a
mistake from copy&paste.

Fixes: ca27d79383129014522ae3278bba269b1fd99ae7

+++ python/addrxlat.c
@@ -105,7 +105,7 @@ PyDoc_STRVAR(attr_convert__doc__,
 static long
 Number_AsLong(PyObject *num)
 {
-	unsigned long long result;
+	long result;
 
 	if (PyLong_Check(num))
 		return PyLong_AsLong(num);
